+++ src/instrumentExplorer.ts
@@ -9,15 +9,16 @@ import { DISCOVER_EXECUTABLE } from "./kic-cli"
 import { LOG_DIR } from "./utility"
 import { ConnectionHelper } from "./resourceManager"
 import { StatusBarManager, StatusType } from "./statusBarManager"
+import { start } from "repl"
 
-const DISCOVERY_TIMEOUT = 5
+const DISCOVERY_TIMEOUT = 4
 
 export class InstrumentsExplorer implements vscode.Disposable {
     private InstrumentsDiscoveryViewer: vscode.TreeView<
         Instrument | Connection | InactiveInstrumentList
     >
     private treeDataProvider?: InstrumentProvider
-    private intervalID?: NodeJS.Timeout
+    private timeoutID?: NodeJS.Timeout
     //private _kicProcessMgr: KicProcessMgr
     private _discoveryInProgress: boolean = false
 
@@ -100,18 +101,19 @@ export class InstrumentsExplorer implements vscode.Disposable {
             StatusType.Progress,
         )
 
-        const interval = setInterval(() => {
+        const start_discovery = () => {
             this.treeDataProvider
                 ?.refresh(async () => await this.startDiscovery())
                 .then(
                     () => {
+                        this.timeoutID = undefined
                         StatusBarManager.instance.hide()
                         if (
-                            !(vscode.workspace
+                            vscode.workspace
                                 .getConfiguration("tsp")
-                                .get("autorefresh") as boolean)
+                                .get("autorefresh") as boolean
                         ) {
-                            clearInterval(interval)
+                            this.timeoutID = setTimeout(start_discovery, 5000)
                         }
                     },
                     (e: Error) => {
@@ -122,7 +124,35 @@ export class InstrumentsExplorer implements vscode.Disposable {
                         )
                     },
                 )
-        }, 5000)
+        }
+        if (
+            vscode.workspace
+                .getConfiguration("tsp")
+                .get("autorefresh") as boolean
+        ) {
+            start_discovery()
+        }
+
+        vscode.workspace.onDidChangeConfiguration((event) => {
+            if (event.affectsConfiguration("tsp.autorefresh")) {
+                if (
+                    !(vscode.workspace
+                        .getConfiguration("tsp")
+                        .get("autorefresh") as boolean) &&
+                    this.timeoutID
+                ) {
+                    clearTimeout(this.timeoutID)
+                    this.timeoutID = undefined
+                } else if (
+                    (vscode.workspace
+                        .getConfiguration("tsp")
+                        .get("autorefresh") as boolean) &&
+                    !this.timeoutID
+                ) {
+                    start_discovery()
+                }
+            }
+        })
     }
     dispose() {
         this.treeDataProvider?.dispose()
@@ -161,7 +191,6 @@ export class InstrumentsExplorer implements vscode.Disposable {
 
                 discover.on("exit", () => {
                     StatusBarManager.instance.hide()
-                    clearInterval(this.intervalID)
                     this._discoveryInProgress = false
                     resolve()
                 })