+++ CHANGELOG.md
@@ -22,8 +22,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Rename 'Upgrade firmware' to 'Update firmware'.
 - Update instrument explorer as new discovered instruments come in rather than after discovery completes
 - **tsp-toolkit-kic-cli** - Run discovery for lan and visa in parallel
+
 ### Added
 - TSP language interop feature.
+- Continuously discover instruments by default (disable by setting the `tsp.autorefresh` setting to `false`)
 - **tsp-toolkit-language-interop** - Implement tsp language interop feature